hydrovision-app/requirements.txt
# HydroVision AG — Backend dependencies
#
# Frontend (CDN, no requiere pip):
#   Tailwind CSS · Alpine.js 3.x · Chart.js 3.9.1 (pinned) · Leaflet 1.9.4
#   Fonts: Inter, IBM Plex Sans, IBM Plex Mono (Google Fonts)

# ── Backend ──────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
jinja2>=3.1.0
python-multipart>=0.0.9   # form parsing (login)

# ── Base de datos ─────────────────────────────────────────────
# PostgreSQL (producción)
psycopg2-binary>=2.9.0

# ── MQTT ─────────────────────────────────────────────────────
paho-mqtt>=1.6.0

# ── Fusión nodo-satélite (Sentinel-2) ────────────────────────
numpy>=1.24.0

# ── Testing ───────────────────────────────────────────────────
pytest>=7.0.0
httpx>=0.24.0      # cliente async para tests de FastAPI
